fix syntax errors, write to output
