+++ latest-release-branch/action.yml
@@ -1,0 +1,33 @@
+name: Get latest release branch
+description: A simple action that returns the latest release branch given a pattern.
+inputs:
+  token:
+    description: GitHub token with access to all necessary repositories
+    required: true
+  ownerRepo:
+    description: Owner/repo which contains the branch (eg `grafana/grafana`)
+    required: true
+  pattern:
+    description: The pattern to use to search for a release branch (eg `v11.3.x`)
+    required: true
+  binary_release_tag:
+    required: false
+    default: "dev"
+outputs:
+  branch:
+    description: The latest release branch that matches the pattern
+runs:
+  using: composite
+  steps:
+  - shell: bash
+    env:
+      GITHUB_TOKEN: ${{inputs.token}}
+      INPUT_OWNERREPO: ${{inputs.ownerRepo}}
+      INPUT_PATTERN: ${{inputs.pattern}}
+    run: |
+      set -e
+      # Download the action from the store
+      curl --fail -L -o /tmp/latest-release-branch https://github.com/grafana/grafana-github-actions-go/releases/download/${{inputs.binary_release_tag}}/latest-release-branch
+      chmod +x /tmp/latest-release-branch
+      # Execute action
+      echo "branch=$(/tmp/latest-release-branch ${{inputs.pattern}})" >> "$GITHUB_OUTPUT"
