+++ pkg/versions/version_branch.go
@@ -8,7 +8,7 @@ import (
 
 func VersionBranch(version string) (string, error) {
 	version = strings.TrimPrefix(version, "v")
-	groups := semverRegexp.FindStringSubmatch(version)
+	groups := SemverRegexp.FindStringSubmatch(version)
 	// The first group is the entire string, so we need 3 results
 	if len(groups) < 3 {
 		return "", errors.New("version does not match a semver regex")
